+++ Cargo.toml
@@ -44,7 +44,6 @@ figment = { version = "0.10.12", features = ["env", "yaml", "test"] }
 flatbuffers = "24.3.25"
 futures = "0.3"
 futures-core = "0.3"
-heck = "0.5.0"
 http-body-util = "0.1.3"
 lazy_static = { version = "1.4" }
 lexical-core = "1.0"
@@ -68,7 +67,6 @@ sea-query-binder = "0.7"
 serde = { version = "1.0.215", features = ["derive"] }
 serde_json = "1.0.133"
 setsum = "0.7"
-syn = "2.0.101"
 tantivy = "0.22.0"
 thiserror = "1.0.69"
 tokio = { version = "1.41", features = ["fs", "macros", "rt-multi-thread"] }
