+++ rust/metering/Cargo.toml
@@ -7,10 +7,8 @@ edition = "2021"
 proc-macro = true
 
 [dependencies]
-heck = { workspace = true }
 proc-macro2 = { workspace = true }
 quote = { workspace = true }
-syn = { workspace = true }
 thiserror = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
